develop : 서비스 계층 Helper 클래스 제거
