+++ src/main/java/dev/linkcentral/common/exception/DomainException.java
@@ -1,0 +1,4 @@
+package dev.linkcentral.common.exception;
+
+public interface DomainException {
+}
