+++ src/main/java/dev/linkcentral/service/FriendService.java
@@ -4,9 +4,9 @@ import dev.linkcentral.database.entity.friend.Friend;
 import dev.linkcentral.database.entity.friend.FriendStatus;
 import dev.linkcentral.database.entity.member.Member;
 import dev.linkcentral.database.repository.friend.FriendRepository;
+import dev.linkcentral.database.repository.member.MemberRepository;
 import dev.linkcentral.service.dto.friend.FriendRequestDTO;
 import dev.linkcentral.service.dto.friend.FriendshipDetailDTO;
-import dev.linkcentral.service.helper.FriendServiceHelper;
 import dev.linkcentral.service.mapper.FriendMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +23,7 @@ public class FriendService {
 
     private final FriendRepository friendRepository;
     private final FriendMapper friendMapper;
-    private final FriendServiceHelper friendServiceHelper;
+    private final MemberRepository memberRepository;
 
     /**
      * 친구 요청을 보냅니다.
@@ -34,9 +34,9 @@ public class FriendService {
      */
     @Transactional
     public Long sendFriendRequest(Long senderId, Long receiverId) {
-        Member sender = friendServiceHelper.findMemberById(senderId);
-        Member receiver = friendServiceHelper.findMemberById(receiverId);
-        friendServiceHelper.validateFriendRequest(sender, receiver);
+        Member sender = findMemberById(senderId);
+        Member receiver = findMemberById(receiverId);
+        validateFriendRequest(sender, receiver);
 
         Friend friendEntity = friendMapper.createFriendRequest(sender, receiver);
         friendRepository.save(friendEntity);
@@ -51,7 +51,7 @@ public class FriendService {
      */
     @Transactional(readOnly = true)
     public List<FriendRequestDTO> getReceivedFriendRequests(Long receiverId) {
-        Member receiver = friendServiceHelper.findMemberById(receiverId);
+        Member receiver = findMemberById(receiverId);
         List<Friend> friends = friendRepository.findAllByReceiverAndStatus(receiver, FriendStatus.REQUESTED);
         return friendMapper.toFriendRequestDTOList(friends);
     }
@@ -63,7 +63,7 @@ public class FriendService {
      */
     @Transactional
     public void acceptFriendRequest(Long requestId) {
-        Friend friendRequest = friendServiceHelper.findFriendRequestByIdAndStatus(requestId, FriendStatus.REQUESTED);
+        Friend friendRequest = findFriendRequestByIdAndStatus(requestId, FriendStatus.REQUESTED);
         friendRequest.updateFriendStatus(FriendStatus.ACCEPTED);
         friendRepository.save(friendRequest);
     }
@@ -75,7 +75,7 @@ public class FriendService {
      */
     @Transactional
     public void rejectFriendRequest(Long requestId) {
-        Friend friendRequest = friendServiceHelper.findFriendRequestById(requestId);
+        Friend friendRequest = findFriendRequestById(requestId);
         friendRepository.delete(friendRequest);
     }
 
@@ -88,8 +88,8 @@ public class FriendService {
      */
     @Transactional
     public Long findFriendshipId(Long senderId, Long receiverId) {
-        Member sender = friendServiceHelper.findMemberById(senderId, "요청자 ID를 찾을 수 없습니다.");
-        Member receiver = friendServiceHelper.findMemberById(receiverId, "대상의 ID를 찾을 수 없습니다.");
+        Member sender = findMemberById(senderId, "요청자 ID를 찾을 수 없습니다.");
+        Member receiver = findMemberById(receiverId, "대상의 ID를 찾을 수 없습니다.");
         return friendRepository.findBySenderAndReceiver(sender, receiver)
                 .map(Friend::getId)
                 .orElseThrow(() -> new EntityNotFoundException("친구 관계 ID를 찾을 수 없습니다."));
@@ -102,7 +102,7 @@ public class FriendService {
      */
     @Transactional
     public void deleteFriendship(Long friendId) {
-        Friend friendship = friendServiceHelper.findFriendRequestById(friendId);
+        Friend friendship = findFriendRequestById(friendId);
         friendRepository.delete(friendship);
     }
 
@@ -113,8 +113,8 @@ public class FriendService {
      */
     @Transactional
     public void unfriendSelected(List<Long> friendIds) {
-        friendServiceHelper.validateFriendIds(friendIds);
-        friendIds.forEach(friendServiceHelper::deleteFriendById);
+        validateFriendIds(friendIds);
+        friendIds.forEach(this::deleteFriendById);
     }
 
     /**
@@ -127,5 +127,85 @@ public class FriendService {
     public List<FriendshipDetailDTO> getFriendships(Long memberId) {
         List<Friend> friendships = friendRepository.findBySenderOrReceiver(memberId);
         return friendMapper.toFriendshipDetailDTOList(friendships);
+    }
+
+    /**
+     * 회원 ID로 회원을 찾습니다.
+     *
+     * @param memberId 회원 ID
+     * @return 회원 엔티티
+     */
+    private Member findMemberById(Long memberId) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new EntityNotFoundException("ID로 멤버를 찾을 수 없습니다."));
+    }
+
+    /**
+     * 회원 ID와 오류 메시지로 회원을 찾습니다.
+     *
+     * @param memberId 회원 ID
+     * @param errorMessage 오류 메시지
+     * @return 회원 엔티티
+     */
+    private Member findMemberById(Long memberId, String errorMessage) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new EntityNotFoundException(errorMessage));
+    }
+
+    /**
+     * 친구 요청을 유효성 검사합니다.
+     *
+     * @param sender 발신자 회원 엔티티
+     * @param receiver 수신자 회원 엔티티
+     */
+    private void validateFriendRequest(Member sender, Member receiver) {
+        boolean friendRequestExists = friendRepository.existsFriendshipBothWays(sender, receiver);
+        if (friendRequestExists) {
+            throw new IllegalStateException("친구 요청을 이미 보냈거나 받았습니다.");
+        }
+    }
+
+    /**
+     * 친구 요청 ID로 친구 요청을 찾습니다.
+     *
+     * @param requestId 친구 요청 ID
+     * @return 친구 엔티티
+     */
+    private Friend findFriendRequestById(Long requestId) {
+        return friendRepository.findById(requestId)
+                .orElseThrow(() -> new EntityNotFoundException("ID로 친구 요청을 찾을 수 없습니다: " + requestId));
+    }
+
+    /**
+     * 친구 요청 ID와 상태로 친구 요청을 찾습니다.
+     *
+     * @param requestId 친구 요청 ID
+     * @param status 친구 요청 상태
+     * @return 친구 엔티티
+     */
+    private Friend findFriendRequestByIdAndStatus(Long requestId, FriendStatus status) {
+        return friendRepository.findByIdAndStatus(requestId, status)
+                .orElseThrow(() -> new EntityNotFoundException("ID로 친구 요청을 찾을 수 없습니다: " + requestId));
+    }
+
+    /**
+     * 친구 ID 목록의 유효성을 검사합니다.
+     *
+     * @param friendIds 친구 ID 목록
+     */
+    private void validateFriendIds(List<Long> friendIds) {
+        if (friendIds == null || friendIds.isEmpty()) {
+            throw new IllegalArgumentException("친구 ID는 null이거나 비워 둘 수 없습니다.");
+        }
+    }
+
+    /**
+     * 친구 ID로 친구 관계를 삭제합니다.
+     *
+     * @param friendId 친구 ID
+     */
+    private void deleteFriendById(Long friendId) {
+        Friend friend = findFriendRequestById(friendId);
+        friendRepository.delete(friend);
     }
 }
