+++ src/main/java/dev/linkcentral/service/StudyMemberService.java
@@ -4,12 +4,14 @@ import dev.linkcentral.database.entity.member.Member;
 import dev.linkcentral.database.entity.studygroup.StudyGroup;
 import dev.linkcentral.database.entity.studygroup.StudyGroupStatus;
 import dev.linkcentral.database.entity.studygroup.StudyMember;
+import dev.linkcentral.database.repository.studygroup.StudyGroupRepository;
+import dev.linkcentral.database.repository.studygroup.StudyMemberRepository;
 import dev.linkcentral.service.dto.studygroup.StudyGroupJoinRequestDTO;
-import dev.linkcentral.service.helper.StudyMemberServiceHelper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -18,7 +20,8 @@ import java.util.stream.Collectors;
 public class StudyMemberService {
 
     private final MemberService memberService;
-    private final StudyMemberServiceHelper studyMemberServiceHelper;
+    private final StudyGroupRepository studyGroupRepository;
+    private final StudyMemberRepository studyMemberRepository;
 
     /**
      * 특정 회원 ID로 스터디 그룹 ID 목록을 가져옵니다.
@@ -28,7 +31,7 @@ public class StudyMemberService {
      */
     @Transactional(readOnly = true)
     public List<Long> getStudyGroupIdsByMemberId(Long memberId) {
-        return studyMemberServiceHelper.findStudyGroupIdsByMemberId(memberId);
+        return studyMemberRepository.findStudyGroupIdsByMemberId(memberId);
     }
 
     /**
@@ -39,9 +42,9 @@ public class StudyMemberService {
     @Transactional
     public void requestJoinStudyGroup(Long studyGroupId) {
         Member currentMember = memberService.getCurrentMember();
-        StudyGroup studyGroup = studyMemberServiceHelper.findStudyGroupById(studyGroupId);
+        StudyGroup studyGroup = findStudyGroupById(studyGroupId);
 
-        studyMemberServiceHelper.validateMemberJoinRequest(currentMember, studyGroup);
+        validateMemberJoinRequest(currentMember, studyGroup);
 
         StudyMember studyMember = StudyMember.builder()
                 .member(currentMember)
@@ -49,7 +52,7 @@ public class StudyMemberService {
                 .status(StudyGroupStatus.REQUESTED)
                 .build();
 
-        studyMemberServiceHelper.saveStudyMember(studyMember);
+        studyMemberRepository.save(studyMember);
     }
 
     /**
@@ -60,9 +63,9 @@ public class StudyMemberService {
      */
     @Transactional(readOnly = true)
     public List<StudyGroupJoinRequestDTO> listJoinRequestsForStudyGroup(Long studyGroupId) {
-        StudyGroup studyGroup = studyMemberServiceHelper.findStudyGroupById(studyGroupId);
+        StudyGroup studyGroup = findStudyGroupById(studyGroupId);
 
-        List<StudyMember> studyMembers = studyMemberServiceHelper.findAllByStudyGroupAndStatus(
+        List<StudyMember> studyMembers = studyMemberRepository.findAllByStudyGroupAndStatus(
                 studyGroup, StudyGroupStatus.REQUESTED);
 
         return studyMembers.stream()
@@ -83,14 +86,14 @@ public class StudyMemberService {
     @Transactional
     public void acceptJoinRequest(Long studyGroupId, Long requestId) {
         Member currentMember = memberService.getCurrentMember();
-        StudyGroup studyGroup = studyMemberServiceHelper.findStudyGroupById(studyGroupId);
+        StudyGroup studyGroup = findStudyGroupById(studyGroupId);
 
-        studyMemberServiceHelper.validateLeaderPermission(currentMember, studyGroup);
+        validateLeaderPermission(currentMember, studyGroup);
 
-        StudyMember joinRequest = studyMemberServiceHelper.findStudyMemberById(requestId);
+        StudyMember joinRequest = findStudyMemberById(requestId);
 
         joinRequest.updateStudyGroupStatus(StudyGroupStatus.ACCEPTED);
-        studyMemberServiceHelper.saveStudyMember(joinRequest);
+        studyMemberRepository.save(joinRequest);
     }
 
     /**
@@ -102,12 +105,12 @@ public class StudyMemberService {
     @Transactional
     public void rejectJoinRequest(Long studyGroupId, Long requestId) {
         Member currentMember = memberService.getCurrentMember();
-        StudyGroup studyGroup = studyMemberServiceHelper.findStudyGroupById(studyGroupId);
+        StudyGroup studyGroup = findStudyGroupById(studyGroupId);
 
-        studyMemberServiceHelper.validateLeaderPermission(currentMember, studyGroup);
+        validateLeaderPermission(currentMember, studyGroup);
 
-        StudyMember joinRequest = studyMemberServiceHelper.findStudyMemberById(requestId);
-        studyMemberServiceHelper.deleteStudyMember(joinRequest);
+        StudyMember joinRequest = findStudyMemberById(requestId);
+        studyMemberRepository.delete(joinRequest);
     }
 
     /**
@@ -117,6 +120,52 @@ public class StudyMemberService {
      */
     @Transactional
     public void removeAllMembersByStudyGroupId(Long studyGroupId) {
-        studyMemberServiceHelper.deleteByStudyGroupId(studyGroupId);
+        studyMemberRepository.deleteByStudyGroupId(studyGroupId);
+    }
+
+    /**
+     * 특정 ID로 스터디 그룹을 찾습니다.
+     *
+     * @param studyGroupId 스터디 그룹 ID
+     * @return StudyGroup 스터디 그룹 엔티티
+     */
+    private StudyGroup findStudyGroupById(Long studyGroupId) {
+        return studyGroupRepository.findById(studyGroupId)
+                .orElseThrow(() -> new EntityNotFoundException("스터디 그룹을 찾을 수 없습니다."));
+    }
+
+    /**
+     * 특정 ID로 스터디 멤버를 찾습니다.
+     *
+     * @param studyMemberId 스터디 멤버 ID
+     * @return StudyMember 스터디 멤버 엔티티
+     */
+    private StudyMember findStudyMemberById(Long studyMemberId) {
+        return studyMemberRepository.findById(studyMemberId)
+                .orElseThrow(() -> new EntityNotFoundException("스터디 회원을 찾을 수 없습니다."));
+    }
+
+    /**
+     * 회원의 스터디 그룹 가입 요청을 유효성 검사합니다.
+     *
+     * @param member 회원 엔티티
+     * @param studyGroup 스터디 그룹 엔티티
+     */
+    private void validateMemberJoinRequest(Member member, StudyGroup studyGroup) {
+        if (studyMemberRepository.existsByMemberAndStudyGroup(member, studyGroup)) {
+            throw new IllegalStateException("이미 스터디 그룹에 지원했습니다.");
+        }
+    }
+
+    /**
+     * 스터디 그룹 리더의 권한을 유효성 검사합니다.
+     *
+     * @param member 회원 엔티티
+     * @param studyGroup 스터디 그룹 엔티티
+     */
+    private void validateLeaderPermission(Member member, StudyGroup studyGroup) {
+        if (!studyGroup.getStudyLeaderId().equals(member.getId())) {
+            throw new IllegalStateException("스터디 그룹 리더만 참여 요청을 수락할 수 있습니다.");
+        }
     }
 }
